# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Promoter API",
        "description": "This API provides CRUD operations for managing promoters",
        "version": "1.0.0"
    },
    "paths": {
        "/api/promoter/{id}": {
            "get": {
                "tags": [
                    "Promoters"
                ],
                "summary": "Get a promoter by ID",
                "description": "Returns a single promoter based on ID",
                "operationId": "getPromoterById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the promoter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promoter"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoter not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Promoters"
                ],
                "summary": "Update a promoter",
                "description": "Update a promoter by ID",
                "operationId": "updatePromoter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the promoter to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the promoter",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "The last name of the promoter",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of the promoter",
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promoter updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoter not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Promoters"
                ],
                "summary": "Delete a promoter",
                "description": "Delete a promoter by ID",
                "operationId": "deletePromoter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the promoter to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promoter deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoter not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/promoter": {
            "post": {
                "tags": [
                    "Promoters"
                ],
                "summary": "Create a new promoter",
                "description": "Creates a new promoter with the provided data",
                "operationId": "storePromoter",
                "requestBody": {
                    "description": "Promoter data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Promoter"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promoter created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating promoter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating promoter"
                                        },
                                        "Errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Error message 1"
                                            },
                                            "example": [
                                                "Error message 1",
                                                "Error message 2"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/promoter/skill-add/{id}": {
            "put": {
                "tags": [
                    "Promoters"
                ],
                "summary": "Add a skill to a promoter",
                "description": "Add a skill to a promoter by ID",
                "operationId": "addSkillToPromoter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the promoter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_id"
                                ],
                                "properties": {
                                    "skill_id": {
                                        "description": "ID of the skill to add to the promoter",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skill added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoter not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promoter not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Promoter": {
                "title": "Promoter",
                "description": "Information about a promoter",
                "properties": {
                    "id": {
                        "description": "The ID of the promoter",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "The name of the promoter",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last Name of the promoter",
                        "type": "string"
                    },
                    "email": {
                        "description": "The Email of the promoter",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The Date of creation of the promoter",
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "description": "Was Updated",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        }
    }
}